refactor: emoji add, more 컴포넌트화
